+++ graphix/pattern.py
@@ -19,13 +19,16 @@ from typing import TYPE_CHECKING, Literal, SupportsFloat, overload
 from warnings import warn
 
 import networkx as nx
-from typing_extensions import assert_never
+
+# override introduced in Python 3.12
+from typing_extensions import assert_never, override
 
 from graphix import command, optimization
 from graphix.command import CommandKind, Node
 from graphix.flow.exceptions import FlowError
 from graphix.fundamentals import Plane
 from graphix.measurements import BlochMeasurement, Measurement, Outcome, toggle_outcome
+from graphix.parameter import InplaceParameterizable
 from graphix.pretty_print import OutputFormat, pattern_to_str
 from graphix.qasm3_exporter import pattern_to_qasm3_lines
 from graphix.sim import DensityMatrix, MBQCTensorNet, Statevector
@@ -69,7 +72,7 @@ class DrawPatternAnnotations(Enum):
     XZCorrections = enum.auto()
 
 
-class Pattern:
+class Pattern(InplaceParameterizable):
     """
     MBQC pattern class.
 
@@ -257,7 +260,7 @@ class Pattern:
             default mapping is computed by using :meth:`node_mapping`.
 
         copy : bool, optional
-            If ``True``, the current pattern remains unchanged and a
+            If ``True``, the current pattern remains unchanged, and a
             new pattern is returned. The default is ``False``, meaning
             that changes are performed in place.
 
@@ -1606,13 +1609,17 @@ class Pattern:
             if cmd.kind == command.CommandKind.M and isinstance(cmd.measurement, BlochMeasurement)
         )
 
-    def subs(self, variable: Parameter, substitute: ExpressionOrSupportsFloat) -> Pattern:
-        """Return a copy of the pattern where all occurrences of the given variable in measurement angles are substituted by the given value."""
-        return self.map(lambda m: m.subs(variable, substitute))
+    @override
+    def replace_parameter(
+        self, variable: Parameter, substitute: ExpressionOrSupportsFloat, *, copy: bool = False
+    ) -> Pattern:
+        return self.apply(lambda m: m.with_parameter(variable, substitute), copy=copy)
 
-    def xreplace(self, assignment: Mapping[Parameter, ExpressionOrSupportsFloat]) -> Pattern:
-        """Return a copy of the pattern where all occurrences of the given keys in measurement angles are substituted by the given values in parallel."""
-        return self.map(lambda m: m.xreplace(assignment))
+    @override
+    def replace_parameters(
+        self, assignment: Mapping[Parameter, ExpressionOrSupportsFloat], *, copy: bool = False
+    ) -> Pattern:
+        return self.apply(lambda m: m.with_parameters(assignment), copy=copy)
 
     def copy(self) -> Pattern:
         """Return a copy of the pattern."""
@@ -1685,8 +1692,47 @@ class Pattern:
                 case CommandKind.C:
                     check_active(cmd, cmd.node)
 
+    def apply(self, f: Callable[[Measurement], Measurement], *, copy: bool = False) -> Pattern:
+        """Apply the function ``f`` to each measurement.
+
+        Parameters
+        ----------
+        f: Callable[[Measurement], Measurement]
+            Function applied to each measurement.
+        copy : bool, optional
+            If ``True``, the current pattern remains unchanged and a
+            new pattern is returned. The default is ``False``, meaning
+            that changes are performed in place.
+
+        Returns
+        -------
+        Pattern
+            The resulting pattern.
+            If ``copy`` is ``False``, the result is ``self``.
+
+        Example
+        -------
+        >>> from graphix import Pattern, command
+        >>> from graphix.measurements import BlochMeasurement, Measurement
+        >>> pattern = Pattern(input_nodes=[0], cmds=[command.M(0, Measurement.XZ(0.25))])
+        >>> pattern.apply(lambda m: BlochMeasurement(m.angle + 1, m.plane))
+        Pattern(input_nodes=[0], cmds=[M(0, Measurement.XZ(1.25))])
+        >>> pattern
+        Pattern(input_nodes=[0], cmds=[M(0, Measurement.XZ(1.25))])
+        """
+        new_seq = [cmd.map(f) if cmd.kind == CommandKind.M else cmd for cmd in self]
+
+        if copy:
+            new_pattern = Pattern(input_nodes=self.input_nodes, cmds=new_seq)
+            new_pattern.reorder_output_nodes(self.output_nodes)
+            return new_pattern
+        self.__seq = new_seq
+        return self
+
     def map(self, f: Callable[[Measurement], Measurement]) -> Pattern:
         """Return a pattern where the function ``f`` has been applied to each measurement.
+
+        Equivalent to ``self.apply(f, copy=True)``.
 
         Parameters
         ----------
@@ -1705,20 +1751,13 @@ class Pattern:
         >>> pattern = Pattern(input_nodes=[0], cmds=[command.M(0, Measurement.XZ(0.25))])
         >>> pattern.map(lambda m: BlochMeasurement(m.angle + 1, m.plane))
         Pattern(input_nodes=[0], cmds=[M(0, Measurement.XZ(1.25))])
+        >>> pattern
+        Pattern(input_nodes=[0], cmds=[M(0, Measurement.XZ(0.25))])
         """
-        new_pattern = Pattern(input_nodes=self.input_nodes)
+        return self.apply(f, copy=True)
 
-        for cmd in self:
-            if cmd.kind == CommandKind.M:
-                new_pattern.add(cmd.map(f))
-            else:
-                new_pattern.add(cmd)
-
-        new_pattern.reorder_output_nodes(self.output_nodes)
-        return new_pattern
-
-    def infer_pauli_measurements(self, rel_tol: float = 1e-09, abs_tol: float = 0.0) -> Pattern:
-        """Return an equivalent pattern in which Bloch measurements close to a Pauli measurement are replaced by Pauli measurements.
+    def infer_pauli_measurements(self, *, rel_tol: float = 1e-09, abs_tol: float = 0.0, copy: bool = False) -> Pattern:
+        """Replace Bloch measurements close to a Pauli measurement by Pauli measurements.
 
         Parameters
         ----------
@@ -1728,11 +1767,17 @@ class Pattern:
         abs_tol : float, optional
             Absolute tolerance for comparing angles, passed to :func:`math.isclose`.
             Default is ``0.0``.
+        copy : bool, optional
+            If ``True``, the current pattern remains unchanged and a
+            new pattern is returned. The default is ``False``, meaning
+            that changes are performed in place.
 
         Returns
         -------
         Pattern
-            An equivalent pattern in which Bloch measurements close to a Pauli measurement are replaced by Pauli measurements.
+            A pattern in which Bloch measurements close to a Pauli
+            measurement are replaced by Pauli measurements. If
+            ``copy`` is ``False``, the result is ``self``.
 
         Example
         -------
@@ -1742,10 +1787,40 @@ class Pattern:
         >>> pattern.infer_pauli_measurements()
         Pattern(input_nodes=[0], cmds=[M(0, Measurement.Y)])
         """
-        return self.map(lambda m: m.to_pauli_or_bloch(rel_tol, abs_tol))
+        return self.apply(lambda m: m.to_pauli_or_bloch(rel_tol, abs_tol), copy=copy)
+
+    def blochify(self, *, copy: bool = False) -> Pattern:
+        """Represent all measurements as Bloch measurements.
+
+        Parameters
+        ----------
+        copy : bool, optional
+            If ``True``, the current pattern remains unchanged and a
+            new pattern is returned. The default is ``False``, meaning
+            that changes are performed in place.
+
+        Returns
+        -------
+        Pattern
+            A pattern in which all measurements are Bloch measurements.
+            If ``copy`` is ``False``, the result is ``self``.
+
+        Example
+        -------
+        >>> from graphix import Pattern, command
+        >>> from graphix.measurements import BlochMeasurement, Measurement
+        >>> pattern = Pattern(input_nodes=[0], cmds=[command.M(0, Measurement.Y)])
+        >>> pattern.blochify()
+        Pattern(input_nodes=[0], cmds=[M(0, Measurement.XY(0.5))])
+        >>> pattern
+        Pattern(input_nodes=[0], cmds=[M(0, Measurement.XY(0.5))])
+        """
+        return self.apply(lambda m: m.to_bloch(), copy=copy)
 
     def to_bloch(self) -> Pattern:
         """Return an equivalent pattern in which all measurements are represented as Bloch measurements.
+
+        Equivalent to ``self.blochify(copy=True)``.
 
         Example
         -------
@@ -1754,15 +1829,17 @@ class Pattern:
         >>> pattern = Pattern(input_nodes=[0], cmds=[command.M(0, Measurement.Y)])
         >>> pattern.to_bloch()
         Pattern(input_nodes=[0], cmds=[M(0, Measurement.XY(0.5))])
+        >>> pattern
+        Pattern(input_nodes=[0], cmds=[M(0, Measurement.Y)])
         """
-        return self.map(lambda m: m.to_bloch())
+        return self.blochify(copy=True)
 
     def perform_pauli_pushing(
         self,
+        *,
         leave_nodes: AbstractSet[Node] | None = None,
         copy: bool = False,
         standardize: bool = False,
-        *,
         stacklevel: int = 1,
     ) -> Pattern:
         """Move Pauli measurements before the other measurements.