+++ tests/test_parameter.py
@@ -3,10 +3,13 @@ from __future__ import annotations
 from typing import TYPE_CHECKING
 
 import matplotlib as mpl
+import networkx as nx
 import numpy as np
 import pytest
 
+# override introduced in Python 3.12
 import graphix.command
+from graphix import OpenGraph
 from graphix.measurements import Measurement
 from graphix.parameter import Placeholder, PlaceholderOperationError
 from graphix.pattern import DrawPatternAnnotations, Pattern
@@ -43,7 +46,7 @@ def test_pattern_without_parameter_subs_is_identity() -> None:
     pattern.add(graphix.command.M(node=0))
     alpha = Placeholder("alpha")
     # Substitution in a pattern without parameterized angle is the identity.
-    assert list(pattern) == list(pattern.subs(alpha, 0))
+    assert list(pattern) == list(pattern.with_parameter(alpha, 0))
 
 
 def test_pattern_substitution() -> None:
@@ -53,11 +56,49 @@ def test_pattern_substitution() -> None:
     pattern.add(graphix.command.M(1, Measurement.XY(alpha)))
     assert pattern.is_parameterized()
     # Parameterized patterns can be substituted, even if some angles are not parameterized.
-    pattern0 = pattern.subs(alpha, 0)
+    pattern0 = pattern.with_parameter(alpha, 0)
     # If all parameterized angles have been instantiated, the pattern is no longer parameterized.
     assert not pattern0.is_parameterized()
     assert list(pattern0) == [graphix.command.M(0), graphix.command.M(1, Measurement.XY(0))]
     assert list(pattern0.infer_pauli_measurements()) == [graphix.command.M(0), graphix.command.M(1)]
+
+
+def test_placeholder_with_parameter_subs() -> None:
+    alpha = Placeholder("alpha")
+    assert alpha.with_parameter(alpha, 1) == 1
+    assert alpha.subs(alpha, 1) == 1
+
+
+def test_placeholder_with_parameters_xreplace() -> None:
+    alpha = Placeholder("alpha")
+    assert alpha.with_parameters({alpha: 1}) == 1
+    assert alpha.xreplace({alpha: 1}) == 1
+
+
+def test_opengraph_with_parameter_subs() -> None:
+    alpha = Placeholder("alpha")
+    og = OpenGraph(graph=nx.Graph([(0, 1)]), input_nodes=[0], output_nodes=[1], measurements={0: Measurement.XY(alpha)})
+    og1 = og.with_parameter(alpha, 0.25)
+    og2 = og.subs(alpha, 0.25)
+    assert og.measurements == {0: Measurement.XY(alpha)}
+    assert og1.measurements == {0: Measurement.XY(0.25)}
+    assert og2.measurements == {0: Measurement.XY(0.25)}
+
+
+def test_opengraph_with_parameters_xreplace() -> None:
+    alpha = Placeholder("alpha")
+    beta = Placeholder("beta")
+    og = OpenGraph(
+        graph=nx.path_graph(3),
+        input_nodes=[0],
+        output_nodes=[2],
+        measurements={0: Measurement.XY(alpha), 1: Measurement.XY(beta)},
+    )
+    og1 = og.with_parameters({alpha: beta, beta: alpha})
+    og2 = og.xreplace({alpha: beta, beta: alpha})
+    assert og.measurements == {0: Measurement.XY(alpha), 1: Measurement.XY(beta)}
+    assert og1.measurements == {0: Measurement.XY(beta), 1: Measurement.XY(alpha)}
+    assert og2.measurements == {0: Measurement.XY(beta), 1: Measurement.XY(alpha)}
 
 
 def test_instantiated_pattern_simulation(fx_rng: Generator) -> None:
@@ -65,10 +106,10 @@ def test_instantiated_pattern_simulation(fx_rng: Generator) -> None:
     pattern.add(graphix.command.M(node=0))
     alpha = Placeholder("alpha")
     pattern.add(graphix.command.M(1, Measurement.XY(alpha)))
-    pattern0 = pattern.subs(alpha, 0)
+    pattern0 = pattern.with_parameter(alpha, 0)
     # Instantied patterns can be simulated.
     pattern0.simulate(rng=fx_rng)
-    pattern1 = pattern.subs(alpha, 1)
+    pattern1 = pattern.with_parameter(alpha, 1)
     assert not pattern1.is_parameterized()
     assert list(pattern1) == [graphix.command.M(node=0), graphix.command.M(1, Measurement.XY(1))]
     assert list(pattern1.infer_pauli_measurements()) == [
@@ -86,9 +127,10 @@ def test_multiple_parameters(fx_rng: Generator) -> None:
     beta = Placeholder("beta")
     pattern.add(graphix.command.N(node=2))
     pattern.add(graphix.command.M(2, Measurement.XY(beta)))
+    pattern.replace_parameter(alpha, 2)
     # A partially instantiated pattern is still parameterized.
-    assert pattern.subs(alpha, 2).is_parameterized()
-    pattern23 = pattern.subs(alpha, 2).subs(beta, 3)
+    assert pattern.is_parameterized()
+    pattern23 = pattern.subs(beta, 3)
     # A full instantiated pattern is no longer parameterized.
     assert not pattern23.is_parameterized()
     assert list(pattern23) == [
@@ -131,22 +173,30 @@ def test_parallel_substitution_with_zero() -> None:
     assert not pattern23.is_parameterized()
 
 
-def test_density_matrix_subs() -> None:
+def test_density_matrix_with_parameter() -> None:
     alpha = Placeholder("alpha")
     dm = DensityMatrix([[alpha]])
-    assert np.allclose(dm.subs(alpha, 1).rho, np.array([1]))
+    assert np.allclose(dm.with_parameter(alpha, 1).rho, np.array([1]))
+    with pytest.raises(TypeError):
+        np.allclose(dm.rho, np.array([1]))
+    dm.replace_parameter(alpha, 1)
+    assert np.allclose(dm.rho, np.array([1]))
 
 
-def test_density_matrix_xreplace() -> None:
+def test_density_matrix_with_parameters() -> None:
     alpha = Placeholder("alpha")
     beta = Placeholder("beta")
     dm = DensityMatrix([[alpha, beta], [alpha, beta]])
-    assert np.allclose(dm.xreplace({alpha: 1, beta: 2}).rho, np.array([[1, 2], [1, 2]]))
+    assert np.allclose(dm.with_parameters({alpha: 1, beta: 2}).rho, np.array([[1, 2], [1, 2]]))
+    with pytest.raises(TypeError):
+        np.allclose(dm.rho, np.array([[1, 2], [1, 2]]))
+    dm.replace_parameters({alpha: 1, beta: 2})
+    assert np.allclose(dm.rho, np.array([[1, 2], [1, 2]]))
 
 
 @pytest.mark.parametrize("jumps", range(1, 11))
-@pytest.mark.parametrize("use_xreplace", [False, True])
-def test_random_circuit_with_parameters(fx_bg: PCG64, jumps: int, use_xreplace: bool) -> None:
+@pytest.mark.parametrize("use_parallel", [False, True])
+def test_random_circuit_with_parameters(fx_bg: PCG64, jumps: int, use_parallel: bool) -> None:
     rng = np.random.Generator(fx_bg.jumped(jumps))
     nqubits = 5
     depth = 5
@@ -156,16 +206,20 @@ def test_random_circuit_with_parameters(fx_bg: PCG64, jumps: int, use_xreplace: 
     pattern = circuit.transpile().pattern
     pattern.standardize()
     pattern.shift_signals()
-    pattern = pattern.infer_pauli_measurements()
+    pattern.infer_pauli_measurements()
     pattern.remove_pauli_measurements()
     pattern.minimize_space()
     assignment: dict[Parameter, float] = {alpha: rng.uniform(high=2), beta: rng.uniform(high=2)}
-    if use_xreplace:
-        state = circuit.xreplace(assignment).simulate().state
-        state_mbqc = pattern.xreplace(assignment).simulate(rng=rng)
+    if use_parallel:
+        circuit.replace_parameters(assignment)
+        pattern.replace_parameters(assignment)
     else:
-        state = circuit.subs(alpha, assignment[alpha]).subs(beta, assignment[beta]).simulate().state
-        state_mbqc = pattern.subs(alpha, assignment[alpha]).subs(beta, assignment[beta]).simulate(rng=rng)
+        circuit.replace_parameter(alpha, assignment[alpha])
+        circuit.replace_parameter(beta, assignment[beta])
+        pattern.replace_parameter(alpha, assignment[alpha])
+        pattern.replace_parameter(beta, assignment[beta])
+    state = circuit.simulate().state
+    state_mbqc = pattern.simulate(rng=rng)
     assert state_mbqc.isclose(state)